requirements.txt
# Tested only with Python 3.4

Django==1.7.1
Pillow==2.6.1
South==1.0.2
argparse==1.3.0
dj-database-url==0.3.0
django-classy-tags==0.5.2
django-cms==3.0.7
django-mptt==0.6.1
django-reversion==1.8.5
django-sekizai==0.7
djangocms-admin-style==0.2.5
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.1
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.6.0
djangocms-link==1.5
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.3
djangocms-video==0.1
djangocms-snippet==1.3
html5lib==0.999
pytz==2014.10
requests==2.5.1
six==1.8.0

#Optional, recommended packages
psycopg2>=2.5
django-bootstrap3
staticsfiles_ignoredebug
django_common_context
django-filer==0.9.8
cmsplugin-filer==0.10.1
easy_thumbnails==2.2
